# internal/api/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Motus API
  version: "0.1.0"
  description: GPS tracking and fleet management API
servers:
  - url: /
    description: This server
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: Session cookie set after login
    bearerAuth:
      type: http
      scheme: bearer
      description: API key token
    xAuthToken:
      type: apiKey
      in: header
      name: X-Auth-Token
      description: iOS PWA fallback — pass session_id value here
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "resource not found"
...
    Command:
      type: object
      required: [id, deviceId, type, status, createdAt]
      properties:
        id: { type: integer, format: int64 }
        deviceId: { type: integer, format: int64 }
        type: { type: string }
        status: { type: string }
        result: { type: string, nullable: true }
        executedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        attributes:
          type: object
          additionalProperties: {}
    CommandInput:
      type: object
      required: [deviceId, type]
      properties:
        deviceId: { type: integer, format: int64 }
        type: { type: string }
        attributes:
          type: object
          additionalProperties: {}
    SendCommandRequest:
      type: object
      required: [deviceId, type]
      properties:
        deviceId: { type: integer, format: int64 }
        type: { type: string }
        attributes:
          type: object
          additionalProperties: {}
    CommandType:
      type: object
      required: [type]
      properties:
        type: { type: string }
...
paths:
...
  /api/commands:
    get:
      operationId: listCommands
      summary: List saved commands
      tags: [commands]
      security:
        - cookieAuth: []
        - bearerAuth: []
        - xAuthToken: []
      parameters:
        - name: deviceId
          in: query
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Command list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Command'
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createCommand
      summary: Save a command (not yet sent)
      tags: [commands]
      security:
        - cookieAuth: []
        - bearerAuth: []
        - xAuthToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommandInput'
      responses:
        "201":
          description: Command created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Command'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /api/commands/send:
    post:
      operationId: sendCommand
      summary: Send a command to a device immediately
      tags: [commands]
      security:
        - cookieAuth: []
        - bearerAuth: []
        - xAuthToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendCommandRequest'
      responses:
        "200":
          description: Command sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Command'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
  /api/commands/types:
    get:
      operationId: getCommandTypes
      summary: List supported command types
      tags: [commands]
      security:
        - cookieAuth: []
        - bearerAuth: []
        - xAuthToken: []
      responses:
        "200":
          description: Command type list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommandType'
        "401":
          $ref: '#/components/responses/Unauthorized'
...